pyproject.toml
[tool.poetry]
name = "hrbot"
version = "0.1.0"
description = "HR Teams Bot with Gemini and RAG"
authors = ["Nedal Altiti <nedala@usclarity.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = "^0.29"
pydantic = "^2.6"
google-cloud-aiplatform = "^1.36.0"
pdfplumber = "^0.10"
python-dotenv = "^1.0"
google-generativeai = "^0.8.5"
httpx = "^0.28.1"
pydantic-settings = "^2.9.1"
python-multipart = "^0.0.20"
tiktoken = "^0.5.2"
faiss-cpu = "^1.7.4"
redis = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"